Fix send queue leaks during CIE and flush completions
CIE completions can arrive even if the send was not
signaled, so these must be discarded when polled.

Additionally, the SQ post_cnt must be decremented
when a flush completion is received, otherwise
the send queue will "fill up" if ENFORCE_SQ_SIZE
is set and a large number of flush completions arrive.

Signed-off-by: Jacob Moroni <jmoroni@google.com>